Move crossfades to SfzHelpers
